+++ src/ts/addons/cargo/cargo.ts
@@ -1,3 +1,5 @@
+import { appendChilds, quickCreate } from "../../util/ui";
+
 export class Item {
     public readonly id: string = "CHANGE_ME";
     public readonly name: string = "Debug Item (YOU SHOULDN'T BE SEEING THIS)";
@@ -30,6 +32,39 @@ export class Cargo {
         });
         
         return items;
+    }
+
+    public cargoloop(rdiv: HTMLDivElement) {
+        rdiv.replaceChildren();
+
+        if (this.cargo.length == 0) {
+            rdiv.appendChild(quickCreate("em", "No ores in the cargo bay."))
+            return;
+        }
+
+        const items = this.getAllItems();
+        const keys = Object.keys(items);
+        
+        keys.forEach(key => {
+            const div = document.createElement("div");
+            // div.id = `${this.id}-cargo-${key}`;
+            div.classList.add("item");
+
+            // todo fix on hover transition
+            div.style.transition = "none";
+
+            const value = items[key];
+
+            const header = document.createElement("h2");
+            const strong = document.createElement("strong");
+            const label = document.createElement("p");
+            header.textContent = value[0];
+            label.textContent = `You have ${value[1]} ${value[0]}`;
+
+            strong?.appendChild(label);
+            appendChilds(div, [header, strong])
+            rdiv.appendChild(div);
+        });
     }
 
     /**
